DockerBackend/requirements.txt
asgiref==3.8.1
config==0.5.1
decouple==0.0.7
Django==4.2.11
djangorestframework==3.15.1
PyMySQL==1.1.0
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
